Implementa módulo ui para las tareas
Este módulo se implementó para agrupar todas las funciones que modifiquen el ui de la página y que tengan que ver con las tareas en sí
